refactor(desktop): run-queue 调度与执行解耦（抽出 PragentRunExecutor）
把 RunQueueService 里「怎么跑一个 run」的逻辑（executeRun + ensureEmbeddedSecrets）整体抽到新的
PragentRunExecutor，调度类只保留并发 / 优先级 / 取消 / 泵：

- 新增 run-queue-types.ts：RunPriority + QueueItem（供调度与执行两侧共享，避免循环依赖）。
- 新增 run-executor.ts：PragentRunExecutor.execute(item, notifyStarted)——落盘 startReviewRun → 物化
  worktree → 组装 env/提示词 → bridge.run → 读产物解析 → finishReviewRun；startedAt 落定经 notifyStarted
  回调通知调度层广播（执行器不持队列态）。
- run-queue.ts 瘦身为纯调度（548 → 186 行）：pump() 改调 this.executor.execute(...)；移除执行相关 import。

公共 API（构造 + enqueue/cancel/snapshot/...）不变，纯搬运、行为不变。desktop typecheck/lint/build 全绿。
