+++ apps/desktop/src/main/services/run-executor.ts
@@ -1,0 +1,364 @@
+import { execFile } from 'node:child_process';
+import fs from 'node:fs/promises';
+import path from 'node:path';
+import { promisify } from 'node:util';
+import { loadAgentRules } from '@meebox/agent';
+import {
+  PRAGENT_LOCAL_OUTPUT,
+  PrAgentRunError,
+  askLanguageSuffixFor,
+  buildExtraInstructions,
+  buildToolEnv,
+  extraInstructionsEnvKey,
+  stripAskQuestionEcho,
+} from '@meebox/pr-agent-bridge';
+import {
+  dropPendingFindingDrafts,
+  finishReviewRun,
+  parseReviewOutput,
+  startReviewRun,
+} from '@meebox/poller';
+import { pickMatchingRule } from '@meebox/rules';
+import type { ReviewRun, ReviewRunStatus } from '@meebox/shared';
+import { getMainLanguage, t } from '../i18n/index.js';
+import { resolveActiveLlmProfile } from '../utils/agent.js';
+import { buildPrContext } from '../utils/pr-context.js';
+import { buildProxyEnv } from '../utils/proxy.js';
+import type { ServiceContext } from './context.js';
+import type { QueueItem } from './run-queue-types.js';
+import {
+  accumulateUsageSentinel,
+  finalizeUsage,
+  newUsageAcc,
+  stripUsageSentinels,
+} from './usage.js';
+
+/**
+ * pr-agent run 的**执行器**（与队列调度 RunQueueService 分离）：给定一个已 dequeue 的队列项，落盘
+ * startReviewRun → 物化 worktree → 组装 env/提示词 → bridge.run → 读产物解析 → finishReviewRun 收尾。
+ * 调度（并发 / 优先级 / 取消 / 泵）归 RunQueueService；本类只管「怎么跑一个 run」，无队列状态。
+ */
+export class PragentRunExecutor {
+  private readonly execFileP = promisify(execFile);
+  /** embedded .secrets.toml 兜底的 memo（只在首个 embedded run 解析一次目录 + 写文件）。 */
+  private embeddedSecretsEnsured: Promise<void> | null = null;
+
+  constructor(private readonly ctx: ServiceContext) {}
+
+  /**
+   * embedded 策略：执行期在嵌入式安装目录的 settings/ 与 settings_prod/ 补空 .secrets.toml
+   * （pr-agent 启动会去找该文件，缺失就打 WARNING；我们走 env 传密钥不用 secrets.toml，写个空
+   * 文件压掉告警）。memo 化：只在首个 embedded run 解析一次目录 + 写文件，后续直接复用。
+   * importlib.util.find_spec 仅定位不 import pr_agent，快；失败仅 warn 不阻断 run。
+   */
+  private ensureEmbeddedSecrets(pythonPath: string): Promise<void> {
+    this.embeddedSecretsEnsured ??= (async () => {
+      const { stdout } = await this.execFileP(pythonPath, [
+        '-c',
+        "import importlib.util,os;print(os.path.dirname(importlib.util.find_spec('pr_agent').origin))",
+      ]);
+      const prAgentDir = stdout.trim();
+      for (const sub of ['settings', 'settings_prod']) {
+        const dir = path.join(prAgentDir, sub);
+        await fs.mkdir(dir, { recursive: true });
+        const f = path.join(dir, '.secrets.toml');
+        try {
+          await fs.access(f);
+        } catch {
+          await fs.writeFile(
+            f,
+            '# meebox placeholder: silence pr-agent warning about a missing .secrets.toml\n',
+          );
+        }
+      }
+    })().catch((err: unknown) => {
+      this.ctx.logger.warn({ err }, 'ensure embedded .secrets.toml failed (ignored)');
+    });
+    return this.embeddedSecretsEnsured;
+  }
+
+  /**
+   * 真正执行一个 queue item：startReviewRun → worktree → bridge.run → finishWith。
+   * 由 RunQueueService.pump() 调用；任何抛错都被调度层兜成 Promise reject，外层 pragent:run 调用方收到。
+   * notifyStarted：startedAt 落定后回调调度层广播队列变化（执行器不持队列态）。
+   */
+  async execute(item: QueueItem, notifyStarted: () => void): Promise<ReviewRun> {
+    const {
+      bootstrap,
+      logger,
+      getPrAgentBridge,
+      embeddedPythonPath,
+      stateStore,
+      repoMirror,
+      broadcast,
+      effectiveAgentDir,
+      pr: prService,
+    } = this.ctx;
+
+    const prAgentBridge = getPrAgentBridge();
+    if (!prAgentBridge) throw new Error(t('prAgent.notReady'));
+    const { req, pr } = item;
+    // 提前 resolve active LLM profile — model 字段要随 startReviewRun 一起落
+    // 盘，让 UI 在 meta 行展示"这次 run 用的什么模型"。后面 buildToolEnv
+    // 同样会用到，这里 resolve 一次复用
+    const activeLlmForRecord = resolveActiveLlmProfile(bootstrap.config.llm);
+    // 用入队预分配的 runId 覆盖 startReviewRun 的自生 id，让 cancel(runId) 在 active
+    // 状态也能精确定位 (跟入队时给的 runId 一致)
+    const run = await startReviewRun(stateStore, {
+      id: item.info.runId,
+      prLocalId: pr.localId,
+      tool: req.tool,
+      question: req.tool === 'ask' ? req.question : undefined,
+      prAgentVersion: prAgentBridge.version,
+      strategy: prAgentBridge.strategy,
+      // 持久化用 profile.model 原文，不做 normalizeModel 前缀处理 — 跟用户
+      // Settings 里看到的名字一致更直观
+      model: activeLlmForRecord?.model || undefined,
+    });
+    // 把入队时 startedAt=null 的 info 升级为 active 形态 + 广播
+    item.info = { ...item.info, startedAt: run.startedAt };
+    notifyStarted();
+    logger.info(
+      { runId: run.id, localId: pr.localId, tool: req.tool, strategy: prAgentBridge.strategy },
+      'pragent run start',
+    );
+    const t0 = Date.now();
+    // 真实 token 用量累加器：sitecustomize 的 litellm callback 把每次调用的 usage 以
+    // `@@MEEBOX_USAGE@@ {json}` 哨兵行打到 stderr，下面 onLine 拦截累加（无需临时文件 / env）。
+    const usageAcc = newUsageAcc();
+    const onLine = (line: string, stream: 'stdout' | 'stderr'): void => {
+      // 拦截 usage 哨兵行：累加后不转发给 renderer（避免污染实时日志）。
+      if (stream === 'stderr' && accumulateUsageSentinel(line, usageAcc)) return;
+      broadcast('pragent:runProgress', { runId: run.id, line, stream });
+    };
+
+    const finishWith = async (patch: Parameters<typeof finishReviewRun>[3]): Promise<ReviewRun> => {
+      const updated = await finishReviewRun(stateStore, pr.localId, run.id, patch);
+      return updated ?? { ...run, ...patch };
+    };
+
+    const repoId = prService.repoIdentityFor(pr);
+    await repoMirror.syncMirror(repoId);
+    // pr-agent 的 LOCAL__TARGET_BRANCH 用固定 merge-base（与 UI diff 同源）：让 AI 评审基于
+    // 「PR 自分叉后引入的改动」，而非 targetRef.sha 漂移后混入别的 PR 的两点对比
+    const diffBase = await prService.resolveDiffBaseSha(pr);
+    const wt = await repoMirror.materializeWorktree(repoId, pr.sourceRef.sha, diffBase);
+    const ac = item.ac!;
+    try {
+      const activeLlm = resolveActiveLlmProfile(bootstrap.config.llm);
+      // 代理 env 先铺底（非 pr-agent 范畴，仅 HTTP(S)_PROXY 类；开关开时让嵌入式 python(litellm/httpx)
+      // 经代理出网调 LLM）；LLM 凭据/模型 + 响应语言 + per-tool pr-agent 配置（/improve 的关项与产出重定向）
+      // 由 bridge 的 buildToolEnv 按意图组装——这些 pr-agent 契约 key 收口在 @meebox/pr-agent-bridge。
+      // 响应语言一期写死在 config 里、UI 还不暴露切换，后续多语言时改成 Settings 入口。
+      const env: Record<string, string> = {
+        ...buildProxyEnv(bootstrap.config.proxy),
+        ...buildToolEnv(activeLlm, { tool: req.tool, responseLanguage: getMainLanguage() }),
+      };
+
+      // PR 上下文 + 命中规则：local provider 不会自己去远端拉，须现读喂给 EXTRA_INSTRUCTIONS；
+      // /ask 跳过（用户问题往往跟历史评论 / 规约无关）。提示词文本的组装见 @meebox/pr-agent-bridge 的 prompts。
+      let prContext = '';
+      let matchedRuleInstructions = '';
+      let matchedRuleId: string | undefined;
+      if (req.tool !== 'ask') {
+        const adapter = prService.adapterFor(pr);
+        if (adapter) {
+          try {
+            prContext = await buildPrContext({ pr, adapter, logger });
+          } catch (err) {
+            logger.warn(
+              { err, runId: run.id, localId: pr.localId },
+              'buildPrContext threw; proceeding without PR context',
+            );
+          }
+        }
+
+        const rules = await loadAgentRules(effectiveAgentDir(), {
+          onWarn: (msg, file) => logger.warn({ file }, `rules: ${msg}`),
+        });
+        const matched = pickMatchingRule(rules, {
+          projectKey: pr.repo.projectKey,
+          repoSlug: pr.repo.repoSlug,
+          targetBranch: pr.targetRef.displayId,
+          tool: req.tool,
+        });
+        if (matched) {
+          matchedRuleInstructions = matched.instructions;
+          matchedRuleId = matched.id;
+        }
+      }
+
+      // 提示词组装收口到 @meebox/pr-agent-bridge 的 prompts：语言指示 / anchor marker / 排版 / PR 上下文 / 命中规则。
+      const extraInstructions = buildExtraInstructions({
+        tool: req.tool,
+        language: getMainLanguage(),
+        prContext,
+        matchedRuleInstructions,
+        // /ask 选中行引用：经 EXTRA_INSTRUCTIONS 注入（不进问题位置参数，故不污染回答 echo）。
+        referencedContext: req.tool === 'ask' ? req.referencedContext : undefined,
+      });
+      if (extraInstructions) env[extraInstructionsEnvKey(req.tool)] = extraInstructions;
+      if (matchedRuleId) {
+        logger.info(
+          { runId: run.id, ruleId: matchedRuleId, tool: req.tool },
+          'pragent run: matched rule',
+        );
+      }
+      if (prContext) {
+        logger.debug(
+          { runId: run.id, tool: req.tool, contextChars: prContext.length },
+          'pragent run: pr context injected',
+        );
+      }
+
+      // ask 工具：问题作为位置参数（user turn，spawn args 单元素，含空格也是一个 arg 不切分），
+      // 并在问题**末尾**硬性追加语言要求。系统侧 CONFIG__RESPONSE_LANGUAGE / EXTRA_INSTRUCTIONS 对
+      // 自由问答常被大量英文 diff（full diff 数万 token）盖过 → 模型用英文作答；在 user turn 末尾
+      // （近因位置、用目标语言书写）再要求一次，显著提升按 UI 语言作答的遵循度。en-US 返回空、不追加。
+      const askLangSuffix = req.tool === 'ask' ? askLanguageSuffixFor(getMainLanguage()) : '';
+      const askQuestion =
+        req.tool === 'ask' && req.question
+          ? askLangSuffix
+            ? `${req.question}\n\n${askLangSuffix}`
+            : req.question
+          : undefined;
+      const extraArgs = askQuestion ? [askQuestion] : undefined;
+
+      // embedded 策略：执行期在嵌入式安装目录补空 .secrets.toml 压掉启动告警
+      // （直接写安装目录；memo 化只首次做）。local-cli 不需要 (pipx 装的 pr-agent
+      // 路径不同，告警也不出)
+      if (prAgentBridge.strategy === 'embedded' && embeddedPythonPath) {
+        await this.ensureEmbeddedSecrets(embeddedPythonPath);
+      }
+
+      const result = await prAgentBridge.run({
+        prUrl: pr.url,
+        tool: req.tool,
+        env,
+        onLine,
+        cwd: wt.path,
+        targetBranch: wt.targetBranchName,
+        extraArgs,
+        signal: ac.signal,
+      });
+      // 真实 token 用量（onLine 累加的 stderr 哨兵行），落到 succeeded / llm-failed 收尾。
+      const tokenUsage = finalizeUsage(usageAcc);
+      // pr-agent 的 local provider 把生成结果写到 worktree 根的 markdown 文件（落盘文件名见 bridge 的
+      // PRAGENT_LOCAL_OUTPUT，与 buildToolEnv 的 LOCAL__REVIEW_PATH 同源）。cleanup 前必须先读出来。
+      const outFile = PRAGENT_LOCAL_OUTPUT[req.tool];
+      let fileContent = '';
+      try {
+        fileContent = await fs.readFile(path.join(wt.path, outFile), 'utf8');
+      } catch (readErr) {
+        logger.warn(
+          { err: readErr, wtPath: wt.path, outFile, runId: run.id },
+          'pr-agent local provider output file missing; fall back to stdout',
+        );
+      }
+      // /ask 输出里 pr-agent 把问题原样回显在 answer body 顶部 (跟 chat 输入气泡完全
+      // 重复)。在解析前把跟用户问题逐字匹配的整行删掉，避免渲染时出现两次问题
+      const cleanedContent =
+        req.tool === 'ask' && req.question?.trim()
+          ? stripAskQuestionEcho(fileContent, req.question, askLangSuffix)
+          : fileContent;
+      const parsed = parseReviewOutput(cleanedContent || result.stdout, req.tool);
+      // M4 草稿再摄入：/review 成功完成时丢掉 pending+finding 旧草稿，
+      // 让本轮 ChatPane 上的 finding 列表成为新的候选源。edited/posted/rejected/
+      // manual 保留不动。失败的 /review 不触发清理 (没建设性数据)。
+      if (req.tool === 'review') {
+        try {
+          const dropped = await dropPendingFindingDrafts(stateStore, pr.localId);
+          if (dropped > 0) {
+            logger.info(
+              { runId: run.id, localId: pr.localId, dropped },
+              'pragent /review: dropped stale pending drafts',
+            );
+            broadcast('drafts:changed', { localId: pr.localId });
+          }
+        } catch (err) {
+          logger.warn({ err, runId: run.id }, 'dropPendingFindingDrafts failed');
+        }
+      }
+      // pr-agent CLI 可能 exit 0 但 stdout 里其实是 LLM 调用全失败 (litellm
+      // AuthenticationError / "Failed to generate prediction with any model" 等
+      // marker)。parseReviewOutput 会在 ParsedReviewOutput.llmFailure 标出 —
+      // 此时不算 succeeded，落盘为 failed + reason='llm-error'，UI 用红色失败
+      // chip 渲染而不是"完成"
+      if (parsed.llmFailure) {
+        logger.warn(
+          { runId: run.id, reason: parsed.llmFailure.message },
+          'pragent exit 0 but LLM call failed; marking run as failed',
+        );
+        return await finishWith({
+          status: 'failed',
+          finishedAt: new Date().toISOString(),
+          durationMs: Date.now() - t0,
+          exitCode: result.exitCode,
+          errorReason: 'llm-error',
+          errorMessage: parsed.llmFailure.message,
+          stdout: fileContent
+            ? `${fileContent}\n\n---\n[pr-agent stdout log]\n${result.stdout}`
+            : result.stdout,
+          stderr: stripUsageSentinels(result.stderr),
+          findings: parsed.findings,
+          summary: parsed.summary,
+          tokenUsage,
+        });
+      }
+      return await finishWith({
+        status: 'succeeded',
+        finishedAt: new Date().toISOString(),
+        durationMs: Date.now() - t0,
+        exitCode: result.exitCode,
+        // 持久化「LLM 真实产出」(文件内容)；stdout 留作日志在折叠区供排障
+        stdout: fileContent
+          ? `${fileContent}\n\n---\n[pr-agent stdout log]\n${result.stdout}`
+          : result.stdout,
+        stderr: stripUsageSentinels(result.stderr),
+        findings: parsed.findings,
+        summary: parsed.summary,
+        tokenUsage,
+      });
+    } catch (err) {
+      if (err instanceof PrAgentRunError) {
+        // 用户主动取消 → status='cancelled'，其它 reason → 'failed'。
+        // 二者都仍走 finishReviewRun 落盘，让 UI 能从历史 run 里看到这次取消事件
+        const status: ReviewRunStatus = err.reason === 'cancelled' ? 'cancelled' : 'failed';
+        logger.warn(
+          { runId: run.id, reason: err.reason, exitCode: err.result.exitCode },
+          `pragent run ${status}`,
+        );
+        // 失败 / 取消时也尽量解析已收集的 stdout：很多情况 pr-agent 已写了一部分输出
+        const partialStdout = err.result.stdout ?? '';
+        const parsed = partialStdout
+          ? parseReviewOutput(partialStdout, req.tool)
+          : { findings: [], summary: undefined };
+        // 失败 / 取消前可能已有若干次 LLM 调用，尽量把已产生的 token 用量也记上
+        const tokenUsage = finalizeUsage(usageAcc);
+        return await finishWith({
+          status,
+          finishedAt: new Date().toISOString(),
+          durationMs: Date.now() - t0,
+          exitCode: err.result.exitCode,
+          errorReason: err.reason,
+          errorMessage: err.message,
+          stdout: err.result.stdout,
+          stderr: stripUsageSentinels(err.result.stderr),
+          findings: parsed.findings,
+          summary: parsed.summary,
+          tokenUsage,
+        });
+      }
+      // 非预期异常：仍记一笔 failed，避免 run 永远卡在 running，再把异常往上抛
+      await finishWith({
+        status: 'failed',
+        finishedAt: new Date().toISOString(),
+        durationMs: Date.now() - t0,
+        errorMessage: err instanceof Error ? err.message : String(err),
+      });
+      throw err;
+    } finally {
+      await wt.cleanup();
+    }
+  }
+}
